get auth headers from client, and set up for multiple users on profile

+++ src/api/app/schemas/profile/resolvers.ts
@@ -1,10 +1,11 @@
-import { ApolloError } from 'apollo-server';
+import { ApolloError, ForbiddenError } from 'apollo-server';
 
 import Profile from './model';
-import UserModel from '../user/model';
+// import UserModel from '../user/model';
 import { ICreateArgs, IQueryArgs, IUpdateArgs, IProfile, IProfileModel} from './types';
 import errorCodes from '../../errorCodes';
-// const auth = require('../../authenticateResolver.js');
+import { IGraphQlContext } from '../../../index';
+import auth from '../../authenticateResolver.js';
 
 // TODO: get error handling dialed
 // https://blog.apollographql.com/full-stack-error-handling-with-graphql-apollo-5c12da407210
@@ -20,52 +21,83 @@ export default {
   },
 
   Mutation: {
-    // async createProfile(root, { input }, { req, res }) {
-      /* try {
-        const user = await auth(req, res);
+    async createProfile(_: {}, { input }: ICreateArgs, { res, req }: IGraphQlContext): Promise<IProfile> {
+      const user = await auth(req, res);
 
-        // does this profile belong to this user???
+      try {
+        // create the new profile
+        const profile = await Profile.create({ 
+          ...input,
+          users: [user],
+        });
 
-        return Profile.create({ ...input, userIds: [user._id] }); // eslint-disable-line
-      } catch (e) {
-        throw new Error(e);
-      }*/
-    async createProfile(_: {}, { input }: ICreateArgs): Promise<IProfile> {
-      const userId = '5bdbe0e075d95e8db4a80bfb';
+        if (!user.profiles) {
+          user.profiles = [];
+        }
 
-      // create the new profile
-      const profile = await Profile.create(input);
+        // save it to the user
+        user.profiles.push(profile);
+        await user.save();
 
-      // save it to the user
-      await UserModel.findOneAndUpdate({ _id: userId }, { $push: { profiles: profile } }, { new: true });
-
-      return profile;
+        return profile;
+      } catch(e) {
+        throw new ApolloError('Failed to create profile', errorCodes.DB_ERROR);
+      }
     },
 
-    updateProfile(_: {}, { id, input }: IUpdateArgs) {
-      return Profile.findOneAndUpdate({ _id: id }, input, {
-        new: true,
-      });
+    async updateProfile(_: {}, { id, input }: IUpdateArgs, { res, req }: IGraphQlContext) {
+      const thisUser = await auth(req, res);
+
+      const profile = await Profile.findOne({ _id: id }).populate('users', ['id']).exec();
+
+      if (!profile) {
+        throw new ApolloError('Profile was not found', errorCodes.NOT_FOUND);
+      }
+
+      if (profile.users.find(user => user.id === thisUser.id)) {
+        profile.set(input);
+        await profile.save();
+
+        return profile;
+      } else {
+        throw new ForbiddenError('This profile does not belong to you');
+      }
     },
 
-    async deleteProfile(_: {}, { id }: IQueryArgs) {
-      const profile = await Profile.findOne({ _id: id });
+    async deleteProfile(_: {}, { id }: IQueryArgs, { res, req }: IGraphQlContext) {
+      const thisUser = await auth(req, res);
+
+      const profile = await Profile.findOne({ _id: id }).populate('users', ['id']).exec();
      
       if (!profile) {
         throw new ApolloError('Profile was not found', errorCodes.NOT_FOUND);
       }
-      
-      await profile.delete();
-      
-      return profile;
+
+      if (profile.users.find(user => user.id === thisUser.id)) {
+        if (profile.users.length < 2) {
+          await profile.delete();
+        } else {
+          // If there are other users, just remove this one
+          // TODO: test this!!!!!!!!
+          profile.set({ $pull: { users: { id: thisUser.id } } });
+
+          await profile.save();
+        }
+
+        return profile;
+      } else {
+        throw new ForbiddenError('This profile does not belong to you');
+      }
     },
 
     async restoreProfile(_: {}, { id }: IQueryArgs) {
-      const profile = await (Profile as any as IProfileModel).findOneDeleted({ _id: id }) as IProfileModel;
+      const profile = await (Profile as any as IProfileModel).findOneWithDeleted({ _id: id }).populate('users', ['id']).exec();
 
       if (!profile) {
         throw new ApolloError('Profile was not found', errorCodes.NOT_FOUND);
       }
+
+      // See if we need to just add the user, or totally restore it ********
 
       await profile.restore();
